Fix GtkBackend/Gtk3Backend openExternalURL compilation error

+++ Sources/Gtk3Backend/Gtk3Backend.swift
@@ -217,7 +217,7 @@ public final class Gtk3Backend: AppBackend {
 
     public func openExternalURL(_ url: URL) throws {
         // Used instead of gtk_uri_launcher_launch to maintain <4.10 compatibility
-        gtk_show_uri(nil, url.absoluteString, GDK_CURRENT_TIME)
+        gtk_show_uri(nil, url.absoluteString, guint(GDK_CURRENT_TIME))
     }
 
     class ThreadActionContext {
