+++ Sources/GtkBackend/GtkBackend.swift
@@ -159,7 +159,7 @@ public final class GtkBackend: AppBackend {
 
     public func openExternalURL(_ url: URL) throws {
         // Used instead of gtk_uri_launcher_launch to maintain <4.10 compatibility
-        gtk_show_uri(nil, url.absoluteString, GDK_CURRENT_TIME)
+        gtk_show_uri(nil, url.absoluteString, guint(GDK_CURRENT_TIME))
     }
 
     private func renderMenu(
